requirements.txt
psycopg2-binary
python-dotenv==0.15.0
pandas

